pyproject.toml
[tool.poetry]
name = "ChattyM - social network"
version = "0.1.0"
description = ""
authors = ["maxx_PC <m.petrykin@gmx.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = ">=5,<6"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.10"
psycopg = "^3.2.10"
drf-spectacular = "^0.28.0"
djangorestframework-simplejwt = "^5.5.1"
django-allauth = {extras = ["google"], version = "^65.11.2"}
dj-rest-auth = "^7.0.1"
pillow = "^11.3.0"
django-widget-tweaks = "^1.5.0"
django-crispy-forms = "^2.4"
crispy-bootstrap5 = "^2025.6"
pytest = "^8.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
